+++ dts-downloader/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "github-webhook-dts-downloader"
+version = "0.1.0"
+edition = "2021"
+
+description = "GitHub webhook payload type def downloader for Rust"
+repository = "https://github.com/sksat/github-webhook-rs"
+authors = ["sksat <sksat@sksat.net>"]
+license = "MIT"
+
+[dependencies]
+github-webhook-type-generator.workspace = true
+anyhow = "1.0.70"
+reqwest = { version = "0.11.16", features = ["blocking"] }
