+++ github-webhook/Cargo.toml
@@ -11,14 +11,9 @@ authors = ["sksat <sksat@sksat.net>"]
 license = "MIT"
 # TODO: how to notice https://github.com/octokit/webhooks/blob/v6.8.0/LICENSE (MIT)
 
-[features]
-default = ["expose-default"]
-expose-default = []
-
 [build-dependencies]
-github-webhook-type-generator.workspace = true
+github-webhook-dts-downloader.workspace = true
 anyhow = "1.0.70"
-reqwest = { version = "0.11.16", features = ["blocking"] }
 
 [dependencies]
 serde = "1.0.160"
