+++ github-webhook/src/lib.rs
@@ -1,2 +1,1 @@
-#[cfg(feature = "expose-default")]
 pub mod payload_types;
